Read Dicom values from HIB and fixed Nifti cutting

+++ org/hipi/image/RasterImage.java
@@ -43,7 +43,7 @@ public abstract class RasterImage extends HipiImage {
 	 * @return color space of image
 	 */
 	public HipiColorSpace getColorSpace() {
-		return HipiColorSpace.fromInteger(header.getValue(HipiImageHeader.JPEG_PNG_INDEX_COLOR_SPACE));
+		return HipiColorSpace.fromInteger((Integer)header.getValue(HipiImageHeader.JPEG_PNG_INDEX_COLOR_SPACE));
 	}
 
 	/**
@@ -52,7 +52,7 @@ public abstract class RasterImage extends HipiImage {
 	 * @return width of image
 	 */
 	public int getWidth() {
-		return header.getValue(HipiImageHeader.JPEG_PNG_INDEX_WIDTH);
+		return (Integer)header.getValue(HipiImageHeader.JPEG_PNG_INDEX_WIDTH);
 	}
 
 	/**
@@ -61,7 +61,7 @@ public abstract class RasterImage extends HipiImage {
 	 * @return height of image
 	 */
 	public int getHeight() {
-		return header.getValue(HipiImageHeader.JPEG_PNG_INDEX_HEIGHT);
+		return (Integer)header.getValue(HipiImageHeader.JPEG_PNG_INDEX_HEIGHT);
 	}
 
 	/**
@@ -70,7 +70,7 @@ public abstract class RasterImage extends HipiImage {
 	 * @return number of color bands in image
 	 */
 	public int getNumBands() {
-		return header.getValue(HipiImageHeader.JPEG_PNG_INDEX_BANDS);
+		return (Integer)header.getValue(HipiImageHeader.JPEG_PNG_INDEX_BANDS);
 	}
 
 	/**
